Add Hybrid Adaptive Voltage Scaling (HAVS) and tweak Static Voltage Scaling (SVS)
The purpose of HAVS is to minimize the power used by the CPU by determining
and setting the optimal voltage. At the same time, the maximum voltage by
which HAVS can scale to is fixed to a specified voltage depending on the
CPU frequency in order to prevent scaling to a higher voltage than what is
normally used at a specified voltage. The optimal voltage is actively determined for each frequency and temperature.